+++ aldryn_forms/admin/base.py
@@ -1,15 +1,28 @@
+import re
 from email.utils import formataddr
+from typing import Callable
+from urllib.parse import urlencode
 
+from django import forms
 from django.conf import settings
-from django.contrib import admin
+from django.contrib import admin, messages
+from django.contrib.sites.models import Site
+from django.db.models.query import QuerySet
+from django.http import HttpRequest, HttpResponse, HttpResponseRedirect, JsonResponse
 from django.template.loader import render_to_string
+from django.template.response import TemplateResponse
 from django.urls import reverse
 from django.utils.encoding import force_str
 from django.utils.module_loading import import_string
+from django.utils.text import slugify
 from django.utils.translation import gettext_lazy as _
 
 from import_export.fields import Field
 from import_export.resources import Resource
+
+from ..api.webhook import collect_submissions_data, send_submissions_data
+from ..models import FormSubmission, Webhook
+from .utils import PrettyJsonEncoder
 
 
 str_dunder_method = '__str__'
@@ -37,8 +50,18 @@ class AldrynFormExportField(Field):
 
 class BaseFormSubmissionAdmin(admin.ModelAdmin):
     date_hierarchy = 'sent_at'
-    list_display = [str_dunder_method, 'sent_at', 'language', 'display_data']
-    list_filter = ['name', 'language']
+    list_display = [
+        str_dunder_method, 'sent_at', 'display_honeypot_filled', 'display_post_ident', 'language', 'display_data'
+    ]
+    list_filter = [
+        'name',
+        'language',
+        'sent_at',
+        ("honeypot_filled", admin.BooleanFieldListFilter),
+        ("post_ident", admin.BooleanFieldListFilter),
+    ]
+    search_fields = ["data"]
+    actions = ["export_webhook", "send_webhook", "honeypot_filled_on", "honeypot_filled_off"]
     readonly_fields = [
         'name',
         'get_data_for_display',
@@ -47,6 +70,16 @@ class BaseFormSubmissionAdmin(admin.ModelAdmin):
         'get_recipients_for_display',
         'post_ident',
         'webhooks',
+    ]
+    fieldsets = [
+        (
+            None,
+            {
+                "fields": [
+                    "name", "get_data_for_display", "language", "sent_at", "webhooks", "form_url", "honeypot_filled"
+                ]
+            }
+        )
     ]
 
     # (Field name, Field label, json data)
@@ -60,6 +93,16 @@ class BaseFormSubmissionAdmin(admin.ModelAdmin):
 
     def has_add_permission(self, request):
         return False
+
+    def get_search_results(self, request, queryset, search_term):
+        queryset, may_have_duplicates = super().get_search_results(request, queryset, search_term)
+        if search_term:
+            try:
+                re.match(search_term, "")
+                queryset |= self.model.objects.filter(data__regex=search_term)
+            except Exception as err:
+                messages.error(request, err)
+        return queryset, may_have_duplicates
 
     def get_data_for_display(self, obj):
         data = obj.get_form_data()
@@ -92,6 +135,8 @@ class BaseFormSubmissionAdmin(admin.ModelAdmin):
 
         url_patterns = [
             path('export/', self.admin_site.admin_view(self.get_form_export_view()), name=self.get_admin_url('export')),
+            path("webhook-export/", self.admin_site.admin_view(self.webhook_export), name="webhook_export"),
+            path("webhook-send/", self.admin_site.admin_view(self.webhook_send), name="webhook_send"),
         ]
         return url_patterns + super(BaseFormSubmissionAdmin, self).get_urls()
 
@@ -203,9 +248,96 @@ class BaseFormSubmissionAdmin(admin.ModelAdmin):
 
         return getattr(resource.export(dataset), export_type)
 
-    def display_data(self, obj):
+    @admin.display(description=_("data"))
+    def display_data(self, obj) -> str:
         if hasattr(settings, 'ALDRYN_FORMS_SUBMISSION_LIST_DISPLAY_FIELD'):
             submission_field = import_string(settings.ALDRYN_FORMS_SUBMISSION_LIST_DISPLAY_FIELD)
             return submission_field(obj)
         return ''
-    display_data.short_description = _('data')
+
+    @admin.display(boolean=True, description=_("Is spam"))
+    def display_honeypot_filled(self, obj) -> bool:
+        return obj.honeypot_filled
+
+    @admin.display(boolean=True, description=_("Ready"))
+    def display_post_ident(self, obj) -> bool:
+        return obj.post_ident is None
+
+    def get_select_webhook_form(self) -> forms.Form:
+        return type("SelectWebhookForm", (forms.Form,), {
+            "webhook": forms.ChoiceField(choices=Webhook.objects.values_list("pk", "name").order_by("name")),
+        })
+
+    def export_submissions_by_webhook(
+        self, request: HttpRequest, submissions: FormSubmission, webhook: Webhook
+    ) -> JsonResponse:
+        site = Site.objects.first()
+        data = collect_submissions_data(webhook, submissions, site.domain)
+        response = JsonResponse({"data": data}, encoder=PrettyJsonEncoder, json_dumps_params={"ensure_ascii": False})
+        filename = f"form-submissions-webhook-{slugify(webhook.name)}.json"
+        response["Content-Disposition"] = f"attachment; filename={filename}"
+        return response
+
+    def send_submissions_data(
+        self, request: HttpRequest, submissions: FormSubmission, webhook: Webhook
+    ) -> HttpResponseRedirect:
+        site = Site.objects.first()
+        send_submissions_data(webhook, submissions, site.domain)
+        messages.success(request, _("Data sending completed."))
+        return HttpResponseRedirect(reverse("admin:aldryn_forms_formsubmission_changelist"))
+
+    def process_webhook(self, request: HttpRequest, process_fnc: Callable, process_title: str) -> HttpResponse:
+        ids = request.GET.get("ids", "")
+        submissions = FormSubmission.objects.filter(pk__in=ids.split("."))
+        SelectWebhookForm = self.get_select_webhook_form()
+        if request.method == "POST":
+            if submissions.count():
+                form = SelectWebhookForm(request.POST)
+                if form.is_valid():
+                    try:
+                        webhook = Webhook.objects.get(pk=form.cleaned_data["webhook"])
+                    except Webhook.DoesNotExist as err:
+                        messages.error(request, err)
+                    else:
+                        return process_fnc(request, submissions, webhook)
+            else:
+                messages.error(request, _("Missing items for processing."))
+            return HttpResponseRedirect(reverse("admin:aldryn_forms_formsubmission_changelist"))
+        else:
+            form = SelectWebhookForm()
+        data = {
+            "ids": ids,
+            "form": form,
+            "submissins_size": submissions.count(),
+            "process_title": process_title,
+        }
+        context = dict(self.admin_site.each_context(request), **data)
+        return TemplateResponse(request, "admin/aldryn_forms/formsubmission/webhook_form.html", context)
+
+    def webhook_export(self, request: HttpRequest) -> HttpResponse:
+        return self.process_webhook(request, self.export_submissions_by_webhook, _("Export data via webhook"))
+
+    def webhook_send(self, request: HttpRequest) -> HttpResponse:
+        return self.process_webhook(request, self.send_submissions_data, _("Send data via webhook"))
+
+    def process_response_redirect(self, queryset: QuerySet, path_name: str) -> HttpResponseRedirect:
+        selected = queryset.values_list("pk", flat=True)
+        params = urlencode({"ids": ".".join([str(pk) for pk in selected])})
+        path = reverse(path_name)
+        return HttpResponseRedirect(f"{path}?{params}")
+
+    @admin.action(description=_("Export data via webhook"), permissions=['change'])
+    def export_webhook(self, request: HttpRequest, queryset: QuerySet) -> HttpResponseRedirect:
+        return self.process_response_redirect(queryset, "admin:webhook_export")
+
+    @admin.action(description=_("Send data via webhook"), permissions=['change'])
+    def send_webhook(self, request: HttpRequest, queryset: QuerySet) -> HttpResponseRedirect:
+        return self.process_response_redirect(queryset, "admin:webhook_send")
+
+    @admin.action(description=_("Set as spam"), permissions=['change'])
+    def honeypot_filled_on(self, request: HttpRequest, queryset: QuerySet) -> HttpResponseRedirect:
+        queryset.update(honeypot_filled=True)
+
+    @admin.action(description=_("Set not to spam"), permissions=['change'])
+    def honeypot_filled_off(self, request: HttpRequest, queryset: QuerySet) -> HttpResponseRedirect:
+        queryset.update(honeypot_filled=False)