+++ aldryn_forms/admin/forms.py
@@ -3,13 +3,18 @@ from datetime import datetime, timedelta
 from django import forms
 from django.conf import settings
 from django.contrib.admin.widgets import AdminDateWidget
+from django.core.exceptions import ValidationError
 from django.utils import timezone
 from django.utils.text import slugify
 from django.utils.translation import gettext
 from django.utils.translation import gettext_lazy as _
 
+from jsonschema import validate
+
+from ..constants import TRANSFORM_SCHEMA
 from ..models import FormSubmission
 from .exporter import Exporter
+from .utils import PrettyJsonEncoder
 
 
 def form_choices(modelClass):
@@ -132,3 +137,17 @@ class FormExportStep2Form(forms.Form):
             message = gettext('Please select at least one field to export.')
             raise forms.ValidationError(message)
         return self.cleaned_data
+
+
+class WebhookAdminForm(forms.ModelForm):
+
+    transform = forms.JSONField(encoder=PrettyJsonEncoder, required=False)
+
+    def clean_transform(self):
+        data = self.cleaned_data["transform"]
+        if data is not None:
+            try:
+                validate(instance=data, schema=TRANSFORM_SCHEMA)
+            except Exception as err:
+                raise ValidationError(err.message)
+        return data