+++ tests/test_api_webhook.py
@@ -9,8 +9,8 @@ from requests.exceptions import HTTPError
 from testfixtures import LogCapture
 
 from aldryn_forms.api.serializers import FormSubmissionSerializer
-from aldryn_forms.api.webhook import send_to_webook, trigger_webhooks
-from aldryn_forms.models import FormSubmission, Webook
+from aldryn_forms.api.webhook import send_to_webhook, trigger_webhooks
+from aldryn_forms.models import FormSubmission, Webhook
 
 
 class Mixin:
@@ -29,7 +29,7 @@ class SendToWebhookTest(Mixin, TestCase):
         with responses.RequestsMock() as rsps:
             rsps.add(responses.POST, self.url, body=HTTPError("Connection failed."))
             with self.assertRaisesMessage(HTTPError, "Connection failed."):
-                send_to_webook(self.url, data)
+                send_to_webhook(self.url, "JSON", data)
         self.log_handler.check()
 
     def test_response(self):
@@ -50,20 +50,21 @@ class SendToWebhookTest(Mixin, TestCase):
         response_data = {"status": "OK"}
         with responses.RequestsMock() as rsps:
             rsps.add(responses.POST, self.url, body=json.dumps(response_data))
-            response = send_to_webook(self.url, payload)
+            response = send_to_webhook(self.url, "JSON", serializer.data)
         self.assertEqual(response.json(), response_data)
         self.assertJSONEqual(payload, data)
         self.log_handler.check()
 
 
+@freeze_time(datetime(2025, 3, 13, 8, 10, tzinfo=timezone.utc))
 class TriggerWebhookTest(Mixin, TestCase):
 
     def setUp(self):
         super().setUp()
-        Webook.objects.create(name="Test", url=self.url)
+        Webhook.objects.create(name="Test", url=self.url)
 
     def test_connection_failed(self):
-        webhooks = Webook.objects.all()
+        webhooks = Webhook.objects.all()
         data = json.dumps([
             {"label": "Test", "name": "test", "value": 1},
         ])
@@ -72,11 +73,15 @@ class TriggerWebhookTest(Mixin, TestCase):
             rsps.add(responses.POST, self.url, body=HTTPError("Connection failed."))
             trigger_webhooks(webhooks, submission, "testserver")
         self.log_handler.check(
+            ('aldryn_forms.api.webhook', 'DEBUG',
+            "{'hostname': 'testserver', 'name': 'Test', 'language': 'en', 'sent_at': "
+            "'2025-03-13T03:10:00-05:00', 'form_recipients': [], 'form_data': "
+            "[{'name': 'test', 'label': 'Test', 'field_occurrence': 1, 'value': 1}]}"),
             ('aldryn_forms.api.webhook', 'ERROR', 'https://host.foo/webhook/ Connection failed.')
         )
 
     def test(self):
-        webhooks = Webook.objects.all()
+        webhooks = Webhook.objects.all()
         data = json.dumps([
             {"label": "Test", "name": "test", "value": 1},
         ])
@@ -84,4 +89,9 @@ class TriggerWebhookTest(Mixin, TestCase):
         with responses.RequestsMock() as rsps:
             rsps.add(responses.POST, self.url, body=json.dumps([{"status": "OK"}]))
             trigger_webhooks(webhooks, submission, "testserver")
-        self.log_handler.check()
+        self.log_handler.check((
+            'aldryn_forms.api.webhook', 'DEBUG',
+            "{'hostname': 'testserver', 'name': 'Test', 'language': 'en', 'sent_at': "
+            "'2025-03-13T03:10:00-05:00', 'form_recipients': [], 'form_data': "
+            "[{'name': 'test', 'label': 'Test', 'field_occurrence': 1, 'value': 1}]}"
+        ))