+++ tests/test_cms_plugins.py
@@ -1,4 +1,5 @@
 import json
+from datetime import datetime, timezone
 from unittest.mock import patch
 
 from django.contrib.auth.models import User
@@ -9,10 +10,11 @@ from cms.api import add_plugin, create_page
 from cms.test_utils.testcases import CMSTestCase
 
 import responses
+from freezegun import freeze_time
 from requests.exceptions import HTTPError
 from testfixtures import LogCapture
 
-from aldryn_forms.models import FormPlugin, FormSubmission, SubmittedToBeSent, Webook
+from aldryn_forms.models import FormPlugin, FormSubmission, SubmittedToBeSent, Webhook
 from tests.test_views import CMS_3_6
 
 
@@ -38,7 +40,7 @@ class DataMixin:
         add_plugin(self.placeholder, 'SubmitButton', 'en', target=self.form_plugin)
         # Webhook
         self.url = "https://host.foo/webhook/"
-        self.webook = Webook.objects.create(name="Test", url=self.url)
+        self.webhook = Webhook.objects.create(name="Test", url=self.url)
         self.log_handler = LogCapture()
         self.addCleanup(self.log_handler.uninstall)
 
@@ -54,6 +56,7 @@ class DataMixin:
             part_html.get_payload())
 
 
+@freeze_time(datetime(2025, 3, 13, 8, 10, tzinfo=timezone.utc))
 class FormPluginTestCase(DataMixin, CMSTestCase):
 
     plugin_name = "FormPlugin"
@@ -125,7 +128,7 @@ class FormPluginTestCase(DataMixin, CMSTestCase):
             self.page.publish('en')
 
         form_plugin = FormPlugin.objects.last()
-        form_plugin.webhooks.add(self.webook)
+        form_plugin.webhooks.add(self.webhook)
         data = {"language": "en", "form_plugin_id": form_plugin.pk, "name": "Tester"}
         with responses.RequestsMock() as rsps:
             rsps.add(responses.POST, self.url, body=HTTPError("Connection failed."))
@@ -138,6 +141,11 @@ class FormPluginTestCase(DataMixin, CMSTestCase):
         ], transform=None)
         self._check_mailbox()
         self.log_handler.check(
+            ('aldryn_forms.api.webhook', 'DEBUG',
+            "{'hostname': 'example.com', 'name': 'Contact us', 'language': 'en', "
+            "'sent_at': '2025-03-13T03:10:00-05:00', 'form_recipients': [{'name': '', "
+            "'email': 'email@example.com'}], 'form_data': [{'name': 'name', 'label': "
+            "'Name', 'field_occurrence': 1, 'value': 'Tester'}]}"),
             ('aldryn_forms.api.webhook', 'ERROR', 'https://host.foo/webhook/ Connection failed.')
         )
 
@@ -148,7 +156,7 @@ class FormPluginTestCase(DataMixin, CMSTestCase):
             self.page.publish('en')
 
         form_plugin = FormPlugin.objects.last()
-        form_plugin.webhooks.add(self.webook)
+        form_plugin.webhooks.add(self.webhook)
         data = {"language": "en", "form_plugin_id": form_plugin.pk, "name": "Tester"}
         with responses.RequestsMock() as rsps:
             rsps.add(responses.POST, self.url, body=json.dumps([{"status": "OK"}]))
@@ -160,7 +168,13 @@ class FormPluginTestCase(DataMixin, CMSTestCase):
             ('Contact us', '[{"name": "name", "label": "Name", "field_occurrence": 1, "value": "Tester"}]', None),
         ], transform=None)
         self._check_mailbox()
-        self.log_handler.check()
+        self.log_handler.check(
+            ('aldryn_forms.api.webhook', 'DEBUG',
+            "{'hostname': 'example.com', 'name': 'Contact us', 'language': 'en', "
+            "'sent_at': '2025-03-13T03:10:00-05:00', 'form_recipients': [{'name': '', "
+            "'email': 'email@example.com'}], 'form_data': [{'name': 'name', 'label': "
+            "'Name', 'field_occurrence': 1, 'value': 'Tester'}]}")
+        )
 
     def test_form_submission_email_action_webhook(self):
         self.form_plugin.action_backend = 'email_only'
@@ -169,7 +183,7 @@ class FormPluginTestCase(DataMixin, CMSTestCase):
             self.page.publish('en')
 
         form_plugin = FormPlugin.objects.last()
-        form_plugin.webhooks.add(self.webook)
+        form_plugin.webhooks.add(self.webhook)
         data = {"language": "en", "form_plugin_id": form_plugin.pk, "name": "Tester"}
         with responses.RequestsMock() as rsps:
             rsps.add(responses.POST, self.url, body=json.dumps([{"status": "OK"}]))
@@ -180,6 +194,9 @@ class FormPluginTestCase(DataMixin, CMSTestCase):
         self._check_mailbox()
         self.log_handler.check(
             ('aldryn_forms.action_backends', 'INFO', 'Sent email notifications to 1 recipients.'),
+            ('aldryn_forms.api.webhook', 'DEBUG',
+            "{'hostname': 'example.com', 'name': 'Contact us', 'language': 'en', "
+            "'sent_at': None, 'form_recipients': [], 'form_data': []}"),
         )
 
     def test_form_submission_no_action_webhook(self):
@@ -189,7 +206,7 @@ class FormPluginTestCase(DataMixin, CMSTestCase):
             self.page.publish('en')
 
         form_plugin = FormPlugin.objects.last()
-        form_plugin.webhooks.add(self.webook)
+        form_plugin.webhooks.add(self.webhook)
         data = {"language": "en", "form_plugin_id": form_plugin.pk, "name": "Tester"}
         with responses.RequestsMock():
             response = self.client.post(self.page.get_absolute_url('en'), data)
@@ -211,7 +228,7 @@ class FormPluginTestCase(DataMixin, CMSTestCase):
             self.page.publish('en')
 
         form_plugin = FormPlugin.objects.last()
-        form_plugin.webhooks.add(self.webook)
+        form_plugin.webhooks.add(self.webhook)
         data = {"language": "en", "form_plugin_id": form_plugin.pk, "name": "Tester"}
         with responses.RequestsMock():
             response = self.client.post(self.page.get_absolute_url('en'), data)
@@ -237,7 +254,7 @@ class FormPluginTestCase(DataMixin, CMSTestCase):
             self.page.publish('en')
 
         form_plugin = FormPlugin.objects.last()
-        form_plugin.webhooks.add(self.webook)
+        form_plugin.webhooks.add(self.webhook)
         data = {"language": "en", "form_plugin_id": form_plugin.pk, "name": "Tester"}
         with responses.RequestsMock():
             response = self.client.post(self.page.get_absolute_url('en'), data)
@@ -259,7 +276,7 @@ class FormPluginTestCase(DataMixin, CMSTestCase):
             self.page.publish('en')
 
         form_plugin = FormPlugin.objects.last()
-        form_plugin.webhooks.add(self.webook)
+        form_plugin.webhooks.add(self.webhook)
         data = {"language": "en", "form_plugin_id": form_plugin.pk, "name": "Tester"}
         with responses.RequestsMock():
             response = self.client.post(self.page.get_absolute_url('en'), data)
@@ -281,7 +298,7 @@ class FormPluginTestCase(DataMixin, CMSTestCase):
             self.page.publish('en')
 
         form_plugin = FormPlugin.objects.last()
-        form_plugin.webhooks.add(self.webook)
+        form_plugin.webhooks.add(self.webhook)
 
         data = {"language": "en", "form_plugin_id": form_plugin.pk, "name": "Tester"}
         with responses.RequestsMock() as rsps:
@@ -300,7 +317,14 @@ class FormPluginTestCase(DataMixin, CMSTestCase):
         msg = mail.outbox[0].message()
         self.assertEqual(msg.get("to"), "email@example.com")
         self.assertEqual(msg.get("subject"), "[Form submission] Contact us")
-        self.log_handler.check()
+        self.log_handler.check(
+            ('aldryn_forms.api.webhook', 'DEBUG',
+            "{'hostname': 'example.com', 'name': 'Contact us', 'language': 'en', "
+            "'sent_at': '2025-03-13T03:10:00-05:00', 'form_recipients': [{'name': '', "
+            "'email': 'email@example.com'}], 'form_data': [{'name': 'name', 'label': "
+            "'Name', 'field_occurrence': 1, 'value': 'Tester'}, {'name': 'trap', "
+            "'label': 'Trap', 'field_occurrence': 1, 'value': ''}]}")
+        )
 
     def test_honeypot_field_filled(self):
         self.form_plugin.action_backend = 'default'
@@ -310,7 +334,7 @@ class FormPluginTestCase(DataMixin, CMSTestCase):
             self.page.publish('en')
 
         form_plugin = FormPlugin.objects.last()
-        form_plugin.webhooks.add(self.webook)
+        form_plugin.webhooks.add(self.webhook)
 
         data = {"language": "en", "form_plugin_id": form_plugin.pk, "name": "Tester", "trap": "Spam!"}
         with responses.RequestsMock():
@@ -324,6 +348,7 @@ class FormPluginTestCase(DataMixin, CMSTestCase):
             'aldryn_forms.cms_plugins', 'INFO', 'Post disabled due to Honeypot "Trap" value: "Spam!"'))
 
 
+@freeze_time(datetime(2025, 3, 13, 8, 10, tzinfo=timezone.utc))
 class EmailNotificationFormPluginTestCase(DataMixin, CMSTestCase):
 
     plugin_name = "EmailNotificationForm"
@@ -395,7 +420,7 @@ class EmailNotificationFormPluginTestCase(DataMixin, CMSTestCase):
             self.page.publish('en')
 
         form_plugin = FormPlugin.objects.last()
-        form_plugin.webhooks.add(self.webook)
+        form_plugin.webhooks.add(self.webhook)
         data = {"language": "en", "form_plugin_id": form_plugin.pk, "name": "Tester"}
         with responses.RequestsMock() as rsps:
             rsps.add(responses.POST, self.url, body=HTTPError("Connection failed."))
@@ -408,6 +433,11 @@ class EmailNotificationFormPluginTestCase(DataMixin, CMSTestCase):
         ], transform=None)
         self._check_mailbox()
         self.log_handler.check(
+            ('aldryn_forms.api.webhook', 'DEBUG',
+            "{'hostname': 'example.com', 'name': 'Contact us', 'language': 'en', "
+            "'sent_at': '2025-03-13T03:10:00-05:00', 'form_recipients': [{'name': '', "
+            "'email': 'email@example.com'}], 'form_data': [{'name': 'name', 'label': "
+            "'Name', 'field_occurrence': 1, 'value': 'Tester'}]}"),
             ('aldryn_forms.api.webhook', 'ERROR', 'https://host.foo/webhook/ Connection failed.')
         )
 
@@ -418,7 +448,7 @@ class EmailNotificationFormPluginTestCase(DataMixin, CMSTestCase):
             self.page.publish('en')
 
         form_plugin = FormPlugin.objects.last()
-        form_plugin.webhooks.add(self.webook)
+        form_plugin.webhooks.add(self.webhook)
         data = {"language": "en", "form_plugin_id": form_plugin.pk, "name": "Tester"}
         with responses.RequestsMock() as rsps:
             rsps.add(responses.POST, self.url, body=json.dumps([{"status": "OK"}]))
@@ -430,7 +460,13 @@ class EmailNotificationFormPluginTestCase(DataMixin, CMSTestCase):
             ('Contact us', '[{"name": "name", "label": "Name", "field_occurrence": 1, "value": "Tester"}]', None),
         ], transform=None)
         self._check_mailbox()
-        self.log_handler.check()
+        self.log_handler.check((
+            'aldryn_forms.api.webhook', 'DEBUG',
+            "{'hostname': 'example.com', 'name': 'Contact us', 'language': 'en', "
+            "'sent_at': '2025-03-13T03:10:00-05:00', 'form_recipients': [{'name': '', "
+            "'email': 'email@example.com'}], 'form_data': [{'name': 'name', 'label': "
+            "'Name', 'field_occurrence': 1, 'value': 'Tester'}]}"
+        ))
 
     def test_form_submission_no_action_webhook(self):
         self.form_plugin.action_backend = 'none'
@@ -439,7 +475,7 @@ class EmailNotificationFormPluginTestCase(DataMixin, CMSTestCase):
             self.page.publish('en')
 
         form_plugin = FormPlugin.objects.last()
-        form_plugin.webhooks.add(self.webook)
+        form_plugin.webhooks.add(self.webhook)
         data = {"language": "en", "form_plugin_id": form_plugin.pk, "name": "Tester"}
         with responses.RequestsMock():
             response = self.client.post(self.page.get_absolute_url('en'), data)